// src/app/page/master/metode-delivery/metode-delivery.interface.ts
import {EventEmitter} from '@angular/core';

export interface MetodeDelivery {
	kode?: any;
	id?: any;
	mataUang?: any;
	kdProfile?: string;
	namaModulAplikasi?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
	namaMetodeDelivery?: string;
}